repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/url_shortener/4/LLMatic/test_admin_dashboard.py
import admin_dashboard as ad
import url_shortener as us
import user_accounts as ua

user_accounts = ua.UserAccounts()


def test_get_all_urls():
	us.generate_short_url('https://example.com')
	assert len(ad.get_all_urls()) >= 1


def test_delete_url():
	short_url = us.generate_short_url('https://example.com')
	assert ad.delete_url(short_url) == 'URL deleted successfully.'
	assert ad.delete_url(short_url) == 'URL does not exist.'


def test_delete_user():
	assert ad.delete_user('testuser') == 'User deleted successfully.'
	assert ad.delete_user('testuser') == 'Username does not exist.'


def test_get_system_performance():
	short_url = us.generate_short_url('https://example.com')
	assert len(ad.get_system_performance()[short_url]) == 0

